+++ lib/screen/personal_account_screen.dart
@@ -3,6 +3,8 @@ import 'package:flutter_svg/flutter_svg.dart';
 
 import '../widgets/my_navigation_bar.dart';
 import '../widgets/outlined_button.dart';
+import '../widgets/playlist_card.dart';
+import '../widgets/tab_bar.dart';
 
 class PersonalAccountScreen extends StatefulWidget {
   const PersonalAccountScreen({Key? key}) : super(key: key);
@@ -116,34 +118,7 @@ class _PersonalAccountScreen extends State<PersonalAccountScreen>
                 ],
               ),
               const SizedBox(height: 10),
-              DecoratedBox(
-                decoration: const BoxDecoration(
-                  border: Border(
-                    bottom: BorderSide(
-                      color: Color(0xFF404663),
-                      width: 2,
-                    ),
-                  ),
-                ),
-                child: TabBar(
-                  indicatorColor: Theme.of(context).colorScheme.onBackground,
-                  labelColor: Theme.of(context).colorScheme.onBackground,
-                  unselectedLabelColor: const Color(0xFF404663),
-                  controller: _tabController,
-                  tabs: const [
-                    Tab(
-                      icon: Icon(
-                        Icons.music_note,
-                      ),
-                    ),
-                    Tab(
-                      icon: Icon(
-                        Icons.history,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
+              TabBars(),
               const SizedBox(height: 25),
               <Widget>[
                 Column(
@@ -183,110 +158,6 @@ class _PersonalAccountScreen extends State<PersonalAccountScreen>
           ),
         ),
       ]),
-    );
-  }
-}
-
-class Label extends StatelessWidget {
-  final String name;
-  final String count;
-
-  const Label({Key? key, required this.name, required this.count})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(5, 0, 0, 10),
-      child: RichText(
-          text: TextSpan(
-        text: '$name ',
-        style: Theme.of(context).textTheme.headlineMedium,
-        children: [
-          const WidgetSpan(
-              alignment: PlaceholderAlignment.baseline,
-              baseline: TextBaseline.alphabetic,
-              child: SizedBox(width: 5)),
-          TextSpan(
-            text: count,
-            style: Theme.of(context).textTheme.headlineMedium?.copyWith(
-                  color: const Color(0xFFA1A9BC),
-                  fontWeight: FontWeight.w400,
-                ),
-          ),
-        ],
-      )),
-    );
-  }
-}
-
-class PlaylistCard extends StatelessWidget {
-  const PlaylistCard({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      child: InkWell(
-          onTap: () {},
-          child: const InfoCard(
-            hasChevron: true,
-          )),
-    );
-  }
-}
-
-class InfoCard extends StatelessWidget {
-  final bool hasChevron;
-
-  const InfoCard({
-    Key? key,
-    this.hasChevron = false,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Padding(
-          padding: const EdgeInsets.all(12),
-          child: ClipRRect(
-            borderRadius: BorderRadius.circular(8.0),
-            child: Image.network(
-              "https://picsum.photos/250?image=9",
-              width: 56,
-              height: 56,
-              fit: BoxFit.cover,
-            ),
-          ),
-        ),
-        Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(
-              "Playlist Name",
-              textAlign: TextAlign.left,
-              style: Theme.of(context)
-                  .textTheme
-                  .displayMedium
-                  ?.copyWith(color: Colors.white),
-            ),
-            const SizedBox(
-              height: 5,
-            ),
-            Text("20 Songs",
-                textAlign: TextAlign.left,
-                style: Theme.of(context).textTheme.displayMedium?.copyWith(
-                      color: const Color(0xFFA1A9BC),
-                    )),
-          ],
-        ),
-        if (hasChevron) const Spacer(),
-        if (hasChevron)
-          const Padding(
-            padding: EdgeInsets.fromLTRB(0, 0, 10, 0),
-            child: Icon(Icons.chevron_right),
-          ),
-      ],
     );
   }
 }
